It's enough to have this as uint32_t, no need for uint64_t

+++ cryptominisat4/gatefinder.h
@@ -264,7 +264,7 @@ private:
     void clean_gates_from_occur();
 
     //For temporaries
-    vector<size_t> seen2Set; //Bits that have been set in seen2, and later need to be cleared
+    vector<uint32_t> seen2Set; //Bits that have been set in seen2, and later need to be cleared
     set<ClOffset> clToUnlink;
     struct TriToUnlink
     {
